storeInfo 변수 추가, idx의 데이터 형 Long으로 변경
StoreInfo 에 당일 예약 가능 여부, 쿠폰 사용 가능 여부에 대한 변수 추가

storeIdx 및 userIdx의 데이터형을 Long으로 변경

매장 조회 시 해당 매장을 방문한 목록에 추가하는 로직을 StoreController에 추가

+++ src/main/java/com/naturemobility/seoul/service/stores/StoreService.java
@@ -12,8 +12,8 @@ import java.util.List;
 public interface StoreService {
 
     public List<SearchStoreRes> retrieveStoreList(String storeName,Double userLatitude,Double userLongitude) throws BaseException;
-    public GetStoreRes retrieveStoreInfoByStoreIdx(Integer storeIdx) throws BaseException;
-    public Integer checkStoreIdx(Integer storeIdx) throws BaseException;
+    public GetStoreRes retrieveStoreInfoByStoreIdx(Long storeIdx) throws BaseException;
+    public Integer checkStoreIdx(Long storeIdx) throws BaseException;
     public List<GetStoreResByMap>retrieveStoreInfoByMap(Double userLatitude, Double userLongitude, Integer orderRule) throws BaseException;
 
 }
